fixed small bugs

+++ lib/functions/attachmentrepository.class.php
@@ -5,8 +5,8 @@
  *
  * Filename $RCSfile: attachmentrepository.class.php,v $
  *
- * @version $Revision: 1.6 $
- * @modified $Date: 2007/12/21 22:57:17 $ by $Author: schlundus $
+ * @version $Revision: 1.7 $
+ * @modified $Date: 2008/01/02 19:34:05 $ by $Author: schlundus $
  * @author Francisco Mancardi
  *
 */
@@ -298,7 +298,8 @@ class tlAttachmentRepository extends tlObjectWithDB
 		if ($bSuccess)
 		{
 			$folder = $this->buildRepositoryFolderFor($fkTableName,$fkid);
-			rmdir($folder);
+			if (is_dir($folder))
+				rmdir($folder);
 		}
 		return $bSuccess;
 	}
